Rename to "Disk Inventory Xs", widen NSUInteger APIs, error-handle NSFileWrapper
User-visible rename + the three PR #3 review followups that weren't in the
dealloc PR.

Rename:
- PRODUCT_NAME, CFBundleName, CFBundleHelpBookName -> "Disk Inventory Xs"
- Localized CFBundleGetInfoString bumped to 1.1 and credited
  "Tjark Derlien 2022, Mahmoud Lababidi 2026"
- All four InfoPlist.strings (en/de/es/fr) replace "Disk Inventory X"
  with "Disk Inventory Xs" in the user-facing permission prompts
- Localized .strings converted from UTF-16 LE to UTF-8 (modern Xcode
  default; makes the diffs reviewable)
- Bundle identifier (com.derlien.DiskInventoryX) and .xcodeproj filename
  intentionally left alone — those break user prefs / external tooling

Public API widening (closes the silent-truncation hole flagged in PR #3
review):
- -[FSItem childCount]: unsigned -> NSUInteger; drop the (unsigned) cast
- -[FSItem childAtIndex:]: unsigned -> NSUInteger to match
- -[FSItem(Utilities) deepFileCountIncludingPackages:]: unsigned -> NSUInteger
- Loop counters at every caller widened from unsigned to NSUInteger
- NSOutlineView delegate methods migrated from int to NSInteger (modern
  Cocoa signature)
- Single narrowing cast retained at the TreeMapView delegate boundary
  because -numberOfChildrenOfItem: in the framework protocol still
  returns unsigned

NSFileWrapper error handling: thread an NSError through the two
-initWithURL:options:error: calls in FSItem.m / NTFilePasteboardSource.m
and skip the pasteboard write if the wrapper failed to load, instead of
serializing a nil wrapper.

BuildRelease.sh:
- Look up the .app's wrapper name from WRAPPER_NAME (it was hard-coded
  to "Disk Inventory X.app")
- Move framework header stripping into BuildRelease.sh after the explicit
  re-sign. The old in-Xcode Run Script phase stripped Headers between
  CodeSignOnCopy and Xcode's bundle CodeSign, leaving the embedded
  framework with a stale signature and failing the .app's sign step.
- Pass CODE_SIGNING_ALLOWED=NO to xcodebuild so Xcode's CodeSign step is
  skipped entirely; BuildRelease.sh's explicit re-sign at the end is the
  only signing pass and produces a consistent signature inner-out.

+++ src/FSItem-Utilities.m
@@ -18,13 +18,13 @@
 
 @implementation FSItem(Utilities)
 
-- (unsigned) deepFileCountIncludingPackages: (BOOL) lookInPackages
+- (NSUInteger) deepFileCountIncludingPackages: (BOOL) lookInPackages
 {
 	if ( [self isSpecialItem] )
 		return 0;
 	
-	unsigned i;
-	unsigned deepChildCount = 0;
+	NSUInteger i;
+	NSUInteger deepChildCount = 0;
 	for ( i = 0; i < [self childCount]; i++ )
 	{
 		FSItem *child = [self childAtIndex: i];
